+++ packages/backend/src/config/lightdashConfig.mock.ts
@@ -149,12 +149,4 @@ export const lightdashConfigMock: LightdashConfig = {
             allowedDomains: [],
         },
     },
-    cube: {
-        token: '',
-    },
-    dbtCloud: {
-        domain: '',
-        bearerToken: '',
-        environmentId: '',
-    },
 };