+++ packages/backend/src/services/UserService.ts
@@ -383,6 +383,18 @@ export class UserService extends BaseService {
             userUuid = existingUserWithEmail.userUuid;
         }
 
+        if (existingUserWithEmail && !existingUserWithEmail?.organizationUuid) {
+            Logger.warn(
+                `User with email ${email} already exists but has no org, so we invite them to join`,
+            );
+            await this.userModel.joinOrg(
+                existingUserWithEmail.userUuid,
+                organizationUuid,
+                userRole,
+                undefined,
+            );
+        }
+
         const inviteLink = await this.inviteLinkModel.upsert(
             inviteCode,
             expiresAt,