+++ packages/frontend/src/hooks/useProject.ts
@@ -4,6 +4,7 @@ import {
     type CreateProject,
     type MostPopularAndRecentlyUpdated,
     type Project,
+    type SemanticLayerConnection,
     type UpdateProject,
 } from '@lightdash/common';
 import { useMutation, useQuery, useQueryClient } from '@tanstack/react-query';
@@ -19,18 +20,28 @@ const createProject = async (data: CreateProject) =>
         body: JSON.stringify(data),
     });
 
-const updateProject = async (id: string, data: UpdateProject) =>
+const updateProject = async (uuid: string, data: UpdateProject) =>
     lightdashApi<ApiJobStartedResults>({
-        url: `/projects/${id}`,
+        url: `/projects/${uuid}`,
         method: 'PATCH',
         body: JSON.stringify(data),
     });
 
-const getProject = async (id: string) =>
+const getProject = async (uuid: string) =>
     lightdashApi<Project>({
-        url: `/projects/${id}`,
+        url: `/projects/${uuid}`,
         method: 'GET',
         body: undefined,
+    });
+
+const updateProjectSemanticLayer = async (
+    uuid: string,
+    data: SemanticLayerConnection,
+) =>
+    lightdashApi<undefined>({
+        url: `/projects/${uuid}/semantic-layer-connection`,
+        method: 'PATCH',
+        body: JSON.stringify(data),
     });
 
 export const useProject = (id: string | undefined) => {
@@ -44,19 +55,19 @@ export const useProject = (id: string | undefined) => {
     });
 };
 
-export const useUpdateMutation = (id: string) => {
+export const useUpdateMutation = (uuid: string) => {
     const queryClient = useQueryClient();
     const { setActiveJobId } = useActiveJob();
     const { showToastApiError } = useToaster();
     return useMutation<ApiJobStartedResults, ApiError, UpdateProject>(
-        (data) => updateProject(id, data),
+        (data) => updateProject(uuid, data),
         {
-            mutationKey: ['project_update', id],
+            mutationKey: ['project_update', uuid],
             onSuccess: async (data) => {
                 setActiveJobId(data.jobUuid);
 
                 await queryClient.invalidateQueries(['projects']);
-                await queryClient.invalidateQueries(['project', id]);
+                await queryClient.invalidateQueries(['project', uuid]);
                 await queryClient.invalidateQueries(['tables']);
                 await queryClient.invalidateQueries(['queryResults']);
                 await queryClient.invalidateQueries(['status']);
@@ -104,3 +115,16 @@ export const useMostPopularAndRecentlyUpdated = (projectUuid: string) =>
         queryKey: ['most-popular-and-recently-updated', projectUuid],
         queryFn: () => getMostPopularAndRecentlyUpdated(projectUuid || ''),
     });
+
+export const useProjectSemanticLayerUpdateMutation = (uuid: string) => {
+    const queryClient = useQueryClient();
+    return useMutation<undefined, ApiError, SemanticLayerConnection>(
+        (data) => updateProjectSemanticLayer(uuid, data),
+        {
+            mutationKey: ['project_semantic_layer_update', uuid],
+            onSuccess: async () => {
+                await queryClient.invalidateQueries(['project', uuid]);
+            },
+        },
+    );
+};